requirements.txt
python-telegram-bot==20.0
pymongo==4.3.3
python-dotenv==0.21.0
spacy==3.4.3
fuzzywuzzy==0.18.0
transformers==4.24.0
tensorflow==2.11.0
